Refactor/clean up front end code (#24)
* Add comments and remove unused imports

* Refactor front end code

* Refactor front end code

* Cleaned up front end code
