+++ react-ui/package.json
@@ -1,45 +1,46 @@
 {
-	"name": "bears-project-3",
-	"version": "0.1.0",
-	"author": "Chingu Winter Beta 2018 Bears 03",
-	"description": "",
-	"keywords": [],
-	"private": true,
-	"main": "",
-	"license": "MIT",
-	"dependencies": {
-		"connected-react-router": "^6.0.0",
-		"history": "^4.7.2",
-		"react": "^16.6.3",
-		"react-dom": "^16.6.3",
-		"react-redux": "^6.0.0",
-		"react-router": "^4.3.1",
-		"react-router-dom": "^4.3.1",
-		"react-router-redux": "^4.0.8",
-		"react-scripts": "2.1.1",
-		"redux": "^4.0.1",
-		"redux-devtools-extension": "^2.13.7",
-		"redux-logger": "^3.0.6",
-		"redux-thunk": "^2.3.0",
-		"semantic-ui-css": "^2.4.1",
-		"semantic-ui-react": "^0.83.0"
-	},
-	"browserslist": [
-		">0.2%",
-		"not dead",
-		"not ie <= 11",
-		"not op_mini all"
-	],
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject"
-	},
-	"devDependencies": {
-		"enzyme": "^3.7.0",
-		"enzyme-adapter-react-16": "^1.7.0",
-		"eslint-config-airbnb": "^17.1.0",
-		"react-hot-loader": "^4.3.12"
-	}
+  "name": "bears-project-3",
+  "version": "0.1.0",
+  "author": "Chingu Winter Beta 2018 Bears 03",
+  "description": "",
+  "keywords": [],
+  "private": true,
+  "main": "./src/index.jsx",
+  "license": "MIT",
+  "dependencies": {
+    "connected-react-router": "^6.0.0",
+    "history": "^4.7.2",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-redux": "^6.0.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "react-router-redux": "^4.0.8",
+    "react-scripts": "2.1.1",
+    "redux": "^4.0.1",
+    "redux-devtools-extension": "^2.13.7",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
+    "semantic-ui-css": "^2.4.1",
+    "semantic-ui-react": "^0.83.0"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ],
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "devDependencies": {
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.7.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "react-hot-loader": "^4.3.12",
+    "prop-types": "^15.6.2"
+  }
 }
