--- react-ui/src/__test__/App.test.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import configureStore from '../store/store';
-import App from '../containers/App/App';
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(
-    <Provider store={configureStore()}>
-      <App />
-    </Provider>, div,
-  );
-  ReactDOM.unmountComponentAtNode(div);
-});
