+++ react-ui/src/components/Auth/presentational/AuthModal.jsx
