+++ react-ui/src/components/Events/presentational/Event.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Card, Icon, Image } from 'semantic-ui-react';
+
+const extra = (event) => {
+  const randomAttendeeCount = () => Math.floor(Math.random() * 15);
+  return (
+    <span style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+      <a href={`localhost:5000/profile:${event.creator.id}`}>
+        <Image
+          style={{ margin: '0 auto' }}
+          src="https://react.semantic-ui.com/images/avatar/large/matthew.png"
+          avatar
+        />
+        <br />
+        {event.creator.name}
+      </a>
+      <a href={`localhost:5000/profile:${event.creator.id}`}>
+        <Icon name="user" />
+        {randomAttendeeCount()}
+        <br />
+        Attendees
+      </a>
+    </span>
+  );
+};
+
+const Event = ({ event }) => (
+  <Card
+    fluid
+    meta={event.category}
+    image="https://react.semantic-ui.com/images/avatar/large/matthew.png"
+    header={event.title}
+    color={event.color}
+    description={event.description}
+    extra={extra(event)}
+  />
+);
+
+Event.propTypes = {
+  event: PropTypes.instanceOf(Object).isRequired,
+};
+
+export default Event;
