+++ react-ui/src/components/HomePage/HomePage.js
@@ -1,0 +1,8 @@
+import { connect } from 'react-redux';
+import HomePage from './presentational/HomePage';
+
+const mapStateToProps = state => ({
+  ...state,
+});
+
+export default connect(mapStateToProps)(HomePage);
