+++ react-ui/src/components/HomePage/presentational/Hero.css
@@ -1,0 +1,3 @@
+h1 {
+  font-family: 'Bungee', cursive;
+}
