--- react-ui/src/containers/HomePage/HomePage.js
@@ -1,8 +1,0 @@
-import { connect } from 'react-redux';
-import HomePage from './components/HomePage';
-
-const mapStateToProps = state => ({
-	...state
-});
-
-export default connect(mapStateToProps)(HomePage);
