--- react-ui/src/containers/HomePage/components/hero.css
@@ -1,3 +1,0 @@
-h1 {
-	font-family: 'Bungee', cursive;
-}
