+++ react-ui/src/store/actions/__tests__/eventsAction.test.js
@@ -1,0 +1,8 @@
+import { SET_FILTER } from '../actionTypes';
+import setFilter from '../eventsAction';
+
+describe('eventsAction', () => {
+  it('should filter events', () => {
+    expect(setFilter('Humor')).toEqual({ type: SET_FILTER, payload: 'Humor' });
+  });
+});
