+++ react-ui/src/store/actions/actionTypes.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const SET_FILTER = 'SET_FILTER';
