+++ react-ui/src/store/actions/eventsAction.js
@@ -1,0 +1,8 @@
+import { SET_FILTER } from './actionTypes';
+
+export default function eventsAction(filter) {
+  return {
+    type: SET_FILTER,
+    payload: filter,
+  };
+}
