--- react-ui/src/store/actions/testAction.js
@@ -1,6 +1,0 @@
-export const testAction = () => (dispatch) => {
-  dispatch({
-    type: 'TEST_ACTION',
-    payload: 'Success! Redux is connected.',
-  });
-};
