+++ react-ui/src/store/eventCategories.js
@@ -1,0 +1,9 @@
+/* eslint-disable import/prefer-default-export */
+export const eventCategories = [
+  'Roleplaying',
+  'Humor',
+  'Adventure',
+  'Action',
+  'Strategy',
+  'Mind',
+];
