--- react-ui/src/store/events/actions.js
@@ -1,10 +1,0 @@
-import { actionTypes as types } from './index';
-
-export const setFilter = filter => ({
-	type: types.SET_FILTER,
-	filter
-});
-
-export const actions = {
-	setFilter
-};
