--- react-ui/src/store/events/index.js
@@ -1,2 +1,0 @@
-export { actions } from './actions';
-export { default as eventsReducer, actionTypes, selectors } from './reducer';
