+++ react-ui/src/store/reducers/rootReducer.js
@@ -2,12 +2,9 @@ import { combineReducers } from 'redux';
 
 import { connectRouter } from 'connected-react-router';
 // Reducers
-import eventsReducer from '../events/reducer';
-import testReducer from './testReducer';
+import eventsReducer from './eventsReducer';
 
-export default history =>
-	combineReducers({
-		events: eventsReducer,
-		test: testReducer,
-		router: connectRouter(history)
-	});
+export default history => combineReducers({
+  events: eventsReducer,
+  router: connectRouter(history),
+});
