--- react-ui/src/store/reducers/testReducer.js
@@ -1,10 +1,0 @@
-export default (state = {}, action) => {
-  switch (action.type) {
-    case 'TEST_ACTION':
-      return {
-        result: action.payload,
-      };
-    default:
-      return state;
-  }
-};
