+++ react-ui/src/store/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import { createBrowserHistory } from 'history';
-import { connectRouter, routerMiddleware } from 'connected-react-router';
+import { routerMiddleware } from 'connected-react-router';
 import { composeWithDevTools } from 'redux-devtools-extension';
 
 // Middleware
@@ -9,15 +9,17 @@ import { createLogger } from 'redux-logger';
 import rootReducer from './reducers/rootReducer';
 
 export const history = createBrowserHistory();
+
 const composeWith = process.env.NODE_ENV === 'development' ? composeWithDevTools : compose;
 const logger = createLogger();
 const enhancers = [];
 
 const middleware = [
-	thunkMiddleware,
-	logger,
-	routerMiddleware(history)
+  thunkMiddleware,
+  logger,
+  routerMiddleware(history),
 ];
+
 const composedEnhancers = composeWith(applyMiddleware(...middleware), ...enhancers);
 
 export default createStore(rootReducer(history), composedEnhancers);
